commit 3/12/2020

+++ proyec_inmobiliariaPilar/apps/persona/views.py
@@ -1,9 +1,11 @@
 from django.shortcuts import render, HttpResponse, redirect
 from .models import PersonaFisica, Persona, PersonaJuridica
 from .forms import PersonaFisicaForm, PersonaJuridicaForm, FiltrarPersonaForm
+from apps.contrato.models import InquilinoPropiedad, PropietarioPropiedad
 import time
 from django.forms import forms
 from django.core.exceptions import ValidationError
+from django.contrib import messages
 
 
 # Create your views here.
@@ -26,10 +28,13 @@ def Nueva_Persona_Fisica(request):
             f.apellido = f.apellido.upper()
             f.desc_per = 'FISICA'
             f.save()
-            data['mensaje']='guardado con exito'
+            messages.success(request,'CLIENTE '+f.nombre+' '+f.apellido+' AGREGADO CORRECTAMENTE')
+            return redirect(to='listado_personas')
+
         data["form"] = formulario
     
     return render(request, "base/nueva_persona.html", data)
+
 
 def Nueva_Persona_Juridica(request):
     data = {
@@ -48,10 +53,12 @@ def Nueva_Persona_Juridica(request):
             f.razon_social = f.razon_social.upper()
             f.desc_per = 'JURIDICA'
             f.save()
-            data['mensaje']='guardado con exito'
+            messages.success(request,'SE AGREGO AL CLIENTE JURIDICO '+f.razon_social+' CORRECTAMENTE')
+            return redirect(to='listado_personas')
         data["form"] = formulario 
     
     return render(request, "base/nueva_persona.html", data)
+
 
 def Modificar_Persona(request, id):
     if PersonaFisica.objects.filter(pk = id).exists():
@@ -67,7 +74,8 @@ def Modificar_Persona(request, id):
                 f.localidad = f.localidad.upper()
                 f.barrio = f.barrio.upper()
                 f.calle = f.calle.upper()
-                f.nombre_apellido = f.nombre_apellido.upper()
+                f.nombre = f.nombre.upper()
+                f.apellido = f.apellido.upper()
                 f.save()
                 data={
                 "mensaje":'guardado correctamente',
@@ -98,23 +106,37 @@ def Modificar_Persona(request, id):
 
     return render(request, "base/nueva_persona.html", data)
 
+
 def Eliminar_Persona(request, id):
     if PersonaFisica.objects.filter(pk = id).exists():
         persona = PersonaFisica.objects.get(pk = id)
+        if InquilinoPropiedad.objects.filter(inquilino = persona, cancelacion = False).exists():
+            messages.error(request, 'NO SE PUEDE ELIMINAR A UN CLIENTE CON UN CONTRATO VIGENTE')
+            return redirect(to='listado_personas')
+        if PropietarioPropiedad.objects.filter(propietario = persona).exists():
+            messages.error(request, 'NO SE PUEDE ELIMINAR A UN CLIENTE CON UN CONTRATO VIGENTE')
+            return redirect(to='listado_personas')
         persona.delete()
+        messages.success(request, 'SE HA ELIMINADO AL CLIENTE ' + persona.nombre + ' ' + persona.apellido+ ' CORRECTAMENTE')
     elif PersonaJuridica.objects.filter(pk = id).exists():
         persona = PersonaJuridica.objects.get(pk= id)
-        persona.delete()     
+        if InquilinoPropiedad.objects.filter(inquilino = persona, cancelacion=False).exists():
+            messages.error(request, 'NO SE PUEDE ELIMINAR A UN CLIENTE CON UN CONTRATO VIGENTE')
+            return redirect(to='listado_personas')
+        if PropietarioPropiedad.objects.filter(propietario = persona).exists():
+            messages.error(request, 'NO SE PUEDE ELIMINAR A UN CLIENTE CON UN CONTRATO VIGENTE')
+            return redirect(to='listado_personas')
+        persona.delete()
+        messages.success(request, 'SE HA ELIMINADO AL CLIENTE ' + persona.razon_social + ' CORRECTAMENTE')     
 
     return redirect(to='listado_personas')
-
 
 
 def Listado_Personas(request):
     data = {
             "form": FiltrarPersonaForm(),
-            "personasFisicas": PersonaFisica.objects.all(),
-            "personasJuridicas":PersonaJuridica.objects.all(),
+            "personas": Persona.objects.all(),
+            #personasJuridicas":PersonaJuridica.objects.all()
     }
     if request.method == 'POST':
         tipo = request.POST.get('tipo')
@@ -123,16 +145,75 @@ def Listado_Personas(request):
                 personasFisicas = PersonaFisica.objects.all()
                 data = {
                     "form": FiltrarPersonaForm(data=request.POST),
-                    "personasFisicas": personasFisicas
+                    "personas": personasFisicas
                 }
             else:
                 personasJuridicas= PersonaJuridica.objects.all()
                 data = {
                     "form": FiltrarPersonaForm(data=request.POST),
-                    "personasJuridicas": personasJuridicas,
+                    "personas": personasJuridicas,
                 }
     
     return render(request, "base/listado_personas.html", data)
 
-def Home(request):
-    return render(request, "base/home.html")
+
+def MostrarPersona(request, id):
+    data = {
+        "id" : "no se puede mostrar"
+    }
+    if PersonaFisica.objects.filter(pk = id).exists():
+        persona = PersonaFisica.objects.get(pk = id)
+        data = {
+            "id" : persona.id,
+            "nombre": persona.nombre + ' ' + persona.apellido,
+            "numero": persona.cuil ,
+            "direccion": persona.provincia + ' - ' + persona.localidad + ' - ' + persona.calle + ' - ' + persona.numero,
+            "telefono": persona.telefono,
+            "mail": persona.mail,
+            "tipo":persona.desc_per
+        }
+    if PersonaJuridica.objects.filter(pk = id).exists():
+        persona = PersonaJuridica.objects.get(pk = id)
+        data = {
+            "id" : persona.id,
+            "nombre": persona.razon_social,
+            "numero": persona.cuit ,
+            "direccion": persona.provincia + ' - ' + persona.localidad + ' - ' + persona.calle + ' - ' + persona.numero,
+            "telefono": persona.telefono,
+            "mail": persona.mail,
+            "tipo": persona.desc_per
+        }
+    
+    
+    return render(request, "base/mostrar_persona.html", data)
+
+
+def ConfirmarPersona(request, id):
+    data = {
+        "id" : "no se puede mostrar"
+    }
+    if PersonaFisica.objects.filter(pk = id).exists():
+        persona = PersonaFisica.objects.get(pk = id)
+        data = {
+            "id" : persona.id,
+            "nombre": persona.nombre + ' ' + persona.apellido,
+            "numero": persona.cuil ,
+            "direccion": persona.provincia + ' - ' + persona.localidad + ' - ' + persona.calle + ' - ' + persona.numero,
+            "telefono": persona.telefono,
+            "mail": persona.mail,
+            "tipo":persona.desc_per
+        }
+    if PersonaJuridica.objects.filter(pk = id).exists():
+        persona = PersonaJuridica.objects.get(pk = id)
+        data = {
+            "id" : persona.id,
+            "nombre": persona.razon_social,
+            "numero": persona.cuit ,
+            "direccion": persona.provincia + ' - ' + persona.localidad + ' - ' + persona.calle + ' - ' + persona.numero,
+            "telefono": persona.telefono,
+            "mail": persona.mail,
+            "tipo": persona.desc_per
+        }
+    
+    
+    return render(request, "base/eliminar_persona.html", data)
